requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
psycopg2-binary
factory-boy
black
isort
flake8
ruff
email-validator
httpx
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
requests
pytest